requirements.txt
numpy>=1.26.0
opencv-python>=4.9.0
pydantic>=2.7.0
PyYAML>=6.0.1
SQLAlchemy>=2.0.29
psycopg2-binary>=2.9.9
fastapi>=0.111.0
uvicorn>=0.30.1
PySide6>=6.7.0
python-dotenv>=1.0.1
pytest>=8.2.0

